pyproject.toml
[tool.poetry]
name = "kenziestarter"
version = "0.1.0"
description = ""
authors = ["Enias Oliveira <eniasoliveira27@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"
Django = "3.1.12"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
